refactor(main): Enhance the functionality of main module
No new features added. But removed several internal functions in the main module as it was not used, such as:
  * `getVideosInfo` (moved to `ytmp3` module)
  * `singleDownload` (moved to `ytmp3` module)
  * `batchDownload` (moved to `ytmp3` module)
  * `normalizeYtMusicUrl` (removed permanently)

In addition, renamed the `argumentParser` function to `getInput`, remove its parameter and small refactor. The function checks whether the given argument is an URL or a file path. If URL is given, the function returns the URL as a `URL` object. Otherwise, if the argument is a file path, the function returns the file path as a string.