feat: Add `DownloadOptions` typedef and its resolver
Added a new typedef called `DownloadOptions` as an interface for the download options object. Later the download options will be used to configure the download process on `singleDownload` and `batchDownload` functions. Additionally, a function to resolve the download options object has been implemented and will strictly check the type for each option.