feat(ytmp3): Add new download method using video ID only
Added a new function called `downloadFromID` which takes a video ID parameter and an optional download options (nullable) to configure the download process. This function will first check if the given ID is a valid YouTube video ID and call the `singleDownload` function to download the YouTube video. Otherwise, if the given ID is invalid an `InvalidTypeError` will be thrown.

By implementing this function, it impacts to the flexibility and simplicity of the project for users.